[APM] Synthtrace improvements
