+++ packages/elastic-apm-synthtrace/scenarios/low_throughput.ts
@@ -7,17 +7,17 @@
  */
 
 import { random } from 'lodash';
-import { apm, timerange } from '../..';
-import { Instance } from '../../lib/apm/instance';
-import { Scenario } from '../scenario';
-import { getLogger } from '../utils/get_common_services';
-import { RunOptions } from '../utils/parse_run_cli_flags';
-import { ApmFields } from '../../lib/apm/apm_fields';
+import { apm, timerange } from '../src';
+import { ApmFields } from '../src/lib/apm/apm_fields';
+import { Instance } from '../src/lib/apm/instance';
+import { Scenario } from '../src/scripts/scenario';
+import { getLogger } from '../src/scripts/utils/get_common_services';
+import { RunOptions } from '../src/scripts/utils/parse_run_cli_flags';
 
 const scenario: Scenario<ApmFields> = async (runOptions: RunOptions) => {
   const logger = getLogger(runOptions);
 
-  const numServices = 500;
+  const numServices = 3;
   const languages = ['go', 'dotnet', 'java', 'python'];
   const services = ['web', 'order-processing', 'api-backend', 'proxy'];
 
@@ -25,7 +25,8 @@ const scenario: Scenario<ApmFields> = async (runOptions: RunOptions) => {
     generate: ({ from, to }) => {
       const range = timerange(from, to);
 
-      const successfulTimestamps = range.interval('1s').rate(3);
+      const successfulTimestamps = range.interval('1s').rate(1);
+      // `.randomize(3, 180);
 
       const instances = [...Array(numServices).keys()].map((index) =>
         apm
@@ -39,15 +40,16 @@ const scenario: Scenario<ApmFields> = async (runOptions: RunOptions) => {
 
       const urls = ['GET /order/{id}', 'POST /basket/{id}', 'DELETE /basket', 'GET /products'];
 
-      const instanceSpans = (instance: Instance, url: string) => {
+      const instanceSpans = (instance: Instance, url: string, index: number) => {
         const successfulTraceEvents = successfulTimestamps.generator((timestamp) => {
-          const randomHigh = random(1000, 4000);
-          const randomLow = random(100, randomHigh / 5);
+          const mod = index % 4;
+          const randomHigh = random(100, mod * 1000);
+          const randomLow = random(10, randomHigh / 10 + mod * 3);
           const duration = random(randomLow, randomHigh);
           const childDuration = random(randomLow, duration);
           const remainderDuration = duration - childDuration;
-          const generateError = random(1, 4) % 3 === 0;
-          const generateChildError = random(0, 5) % 2 === 0;
+          const generateError = index % random(mod, 9) === 0;
+          const generateChildError = index % random(mod, 9) === 0;
           const span = instance
             .transaction(url)
             .timestamp(timestamp)
@@ -77,8 +79,8 @@ const scenario: Scenario<ApmFields> = async (runOptions: RunOptions) => {
 
       return instances
         .flatMap((instance) => urls.map((url) => ({ instance, url })))
-        .map(({ instance, url }) =>
-          logger.perf('generating_apm_events', () => instanceSpans(instance, url))
+        .map(({ instance, url }, index) =>
+          logger.perf('generating_apm_events', () => instanceSpans(instance, url, index))
         )
         .reduce((p, c) => p.merge(c));
     },
