+++ packages/elastic-apm-synthtrace/scenarios/span_links.ts
@@ -7,9 +7,9 @@
  */
 
 import { compact, shuffle } from 'lodash';
-import { apm, ApmFields, EntityArrayIterable, timerange } from '../..';
-import { generateLongId, generateShortId } from '../../lib/utils/generate_id';
-import { Scenario } from '../scenario';
+import { apm, ApmFields, EntityArrayIterable, timerange } from '../src';
+import { generateLongId, generateShortId } from '../src/lib/utils/generate_id';
+import { Scenario } from '../src/scripts/scenario';
 
 function generateExternalSpanLinks() {
   // randomly creates external span links 0 - 10
